Permutation and platform handling adjust, refactored host by removing a bit of duplicate code.
