+++ src/PerformanceTests/Common.Data/Common.Data.projitems
@@ -15,7 +15,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)ConcurrentyLevelConverter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ICreateTestData.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IPermutation.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Log.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)EnvironmentStats.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PerformanceCounters.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Program.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Scenarios\SendLocalOneOnOneRunner.cs" />
