+++ src/PerformanceTests/Common.Data/Program.cs
@@ -2,13 +2,14 @@ namespace Host
 {
     using System;
     using System.Linq;
-    using Common;
+    using NServiceBus.Logging;
     using Tests.Permutations;
     using Utils;
     using VisualStudioDebugHelper;
 
     partial class Program
     {
+        static readonly ILog Log = LogManager.GetLogger(typeof(Program));
         static string endpointName = "PerformanceTests_" + AppDomain.CurrentDomain.FriendlyName.Replace(' ', '_');
         static void Main(string[] args)
         {
@@ -20,12 +21,12 @@ namespace Host
 
                 Statistics.Initialize();
 
-                Log.Env();
+                EnvironmentStats.Write();
 
                 var permutation = PermutationParser.FromCommandlineArgs();
                 var options = BusCreationOptions.Parse(args);
 
-                if (Environment.UserInteractive) Console.Title = permutation.ToString();
+                if (Environment.UserInteractive) Console.Title = PermutationParser.ToString(permutation);
 
                 var assembly = System.Reflection.Assembly.GetExecutingAssembly();
                 var tasks = permutation.Tests.Select(x => (IStartAndStop)assembly.CreateInstance(x)).ToArray();
@@ -34,9 +35,21 @@ namespace Host
             }
             catch (Exception ex)
             {
-                NServiceBus.Logging.LogManager.GetLogger(typeof(Program)).Fatal("Main", ex);
+                Log.Fatal("Main", ex);
                 throw;
             }
+        }
+
+        static void Run(IStartAndStop[] tasks)
+        {
+            foreach (var t in tasks) t.Start();
+            Log.InfoFormat("Warmup");
+            System.Threading.Thread.Sleep(Settings.WarmupDuration);
+            Statistics.Instance.Reset();
+            Log.InfoFormat("Run");
+            System.Threading.Thread.Sleep(Settings.RunDuration);
+            Statistics.Instance.Dump();
+            foreach (var t in tasks) t.Stop();
         }
     }
 }
